backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["LongReeds"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
alembic = "^1.13.2"
sqlmodel = "^0.0.21"
passlib = "^1.7.4"
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
uvicorn = "^0.30.5"
aioboto3 = "^13.1.1"
pyjwt = "^2.9.0"
bcrypt = "4.0.1"
requests = "^2.32.3"
httpx = "^0.23.0"
sqlalchemy-utils = "^0.41.2"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^22.6.0"
ConfigArgParse = "^1.5.3"
isort = "^5.10.1"
mock = "^4.0.3"
mypy = "^0.971"
pylint = "^2.14.5"
pytest = "^7.1.2"
pytest-aiohttp = "^1.0.4"
pytest-cov = "^3.0.0"
sqlalchemy2-stubs = "^0.0.2-alpha.24"
watchfiles = "^0.16.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
python_files = "test_*.py"
python_classes = ["*Test", "Test*"]
python_functions = "test_*"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "# pragma: no cover",
    "def __repr__",
    "def __str__",
    "def __unicode__",
]
show_missing = true
skip_empty = true
omit = [
  "*/__init__.py",
  "shortener/db/migrator/env.py",
]

[tool.coverage.html]
directory = "pytest-cov-report"

[tool.coverage.run]
branch = true
concurrency = ['thread', 'greenlet']

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pylint.master]
extension-pkg-allow-list = ["pydantic"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.basic]
good-names = ["db"]

[tool.pylint."MESSAGES CONTROL"]
disable = [
  "C0114",  # missing-module-docstring
  "C0116",  # missing-function-docstring
  "R0903",  # too-few-public-methods (for sqlalchemy tables and pydantic classes)
  "C0115",  # missing-class-docstring (for api routes)
]

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = false
no_implicit_optional = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
known_local_folder = "bookmarker"
lines_after_imports = 2
multi_line_output = 3
line_length = 120
include_trailing_comma = true
